css/css-color-adjust/meta/meta-color-scheme-attribute-changes.html

<!doctype html>
<title>CSS Color Adjust Test: meta color-scheme - attribute changes</title>
<meta id="meta" name="color-scheme" content="dark">
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<script>
  assert_root_color_scheme("dark");

  meta.removeAttribute("name");
  assert_root_color_scheme("normal");

  meta.setAttribute("name", "color-scheme");
  assert_root_color_scheme("dark");

  meta.setAttribute("content", "");
  assert_root_color_scheme("normal");

  meta.setAttribute("content", ",,invalid");
  assert_root_color_scheme("normal");

  meta.setAttribute("content", "light");
  assert_root_color_scheme("light");

  meta.removeAttribute("content");
  assert_root_color_scheme("normal");
</script>
